Refactoring: functions to work with JSON responses
